+++ crates/ruma-common/src/api.rs
@@ -318,6 +318,10 @@ pub enum SendAccessToken<'a> {
     /// Always add the access token.
     Always(&'a str),
 
+    /// Add the given appservice token to the request only if the `METADATA` on the request
+    /// requires it.
+    Appservice(&'a str),
+
     /// Don't add an access token.
     ///
     /// This will lead to an error if the request endpoint requires authentication
@@ -329,7 +333,7 @@ impl<'a> SendAccessToken<'a> {
     ///
     /// Returns `Some(_)` if `self` contains an access token.
     pub fn get_required_for_endpoint(self) -> Option<&'a str> {
-        as_variant!(self, Self::IfRequired | Self::Always)
+        as_variant!(self, Self::IfRequired | Self::Appservice | Self::Always)
     }
 
     /// Get the access token for an endpoint that should not require one.
@@ -337,6 +341,14 @@ impl<'a> SendAccessToken<'a> {
     /// Returns `Some(_)` only if `self` is `SendAccessToken::Always(_)`.
     pub fn get_not_required_for_endpoint(self) -> Option<&'a str> {
         as_variant!(self, Self::Always)
+    }
+
+    /// Gets the access token for an endpoint that requires one for appservices.
+    ///
+    /// Returns `Some(_)` if `self` is either `SendAccessToken::Appservice(_)`
+    /// or `SendAccessToken::Always(_)`
+    pub fn get_required_for_appservice(self) -> Option<&'a str> {
+        as_variant!(self, Self::Appservice | Self::Always)
     }
 }
 
@@ -489,6 +501,12 @@ pub enum AuthScheme {
     ///
     /// It is recommended to use the header over the query parameter.
     AccessTokenOptional,
+
+    /// Authentication is only performed for appservices, by including an access token in the
+    /// `Authentication` http header, or an `access_token` query parameter.
+    ///
+    /// It is recommended to use the header over the query parameter.
+    AppserviceToken,
 
     /// Authentication is performed by including X-Matrix signatures in the request headers,
     /// as defined in the federation API.